:dizzy: /group/:slug/:id moves to routes/group_thread.js
